Add MenuManager
