+++ lib/main-menu.js
@@ -7,18 +7,8 @@ const macOS = process.platform === 'darwin'
 // const WIN = process.platform === 'win32'
 const LINUX = process.platform === 'linux'
 
-const Config = require('electron-config')
-const config = new Config()
-const DEFAULT_SHORTCUTS = {
-  newNote: 'CommandOrControl+N',
-  focusNote: 'Control+E',
-  deleteNote: macOS? 'Control+Backspace' : 'Control+Delete',
-  print: 'CommandOrControl+P',
-  nextNote: 'Control+J',
-  previousNote: 'Control+K',
-  focusSearch: 'Control+S'
-}
-const menuShortcuts = config.get('menuShortcuts') || DEFAULT_SHORTCUTS
+const MenuManager = require('../browser/main/lib/MenuManager')
+const menuShortcuts = MenuManager.getShortcuts()
 
 const boost = macOS
   ? {
@@ -68,14 +58,14 @@ const file = {
   submenu: [
     {
       label: 'New Note',
-      accelerator: menuShortcuts.newNote || DEFAULT_SHORTCUTS.newNote,
+      accelerator: menuShortcuts.newNote,
       click () {
         mainWindow.webContents.send('top:new-note')
       }
     },
     {
       label: 'Focus Note',
-      accelerator: menuShortcuts.focusNote || DEFAULT_SHORTCUTS.focusNote,
+      accelerator: menuShortcuts.focusNote,
       click () {
         mainWindow.webContents.send('detail:focus')
       }
@@ -121,7 +111,7 @@ const file = {
     },
     {
       label: 'Print',
-      accelerator: menuShortcuts.print || DEFAULT_SHORTCUTS.print,
+      accelerator: menuShortcuts.print,
       click () {
         mainWindow.webContents.send('list:isMarkdownNote')
         mainWindow.webContents.send('print')
@@ -132,7 +122,7 @@ const file = {
     },
     {
       label: 'Delete Note',
-      accelerator: menuShortcuts.deleteNote || DEFAULT_SHORTCUTS.deleteNote,
+      accelerator: menuShortcuts.deleteNote,
       click () {
         mainWindow.webContents.send('detail:delete')
       }
@@ -209,14 +199,14 @@ const view = {
     },
     {
       label: 'Next Note',
-      accelerator: menuShortcuts.nextNote || DEFAULT_SHORTCUTS.nextNote,
+      accelerator: menuShortcuts.nextNote,
       click () {
         mainWindow.webContents.send('list:next')
       }
     },
     {
       label: 'Previous Note',
-      accelerator: menuShortcuts.previousNote || DEFAULT_SHORTCUTS.previousNote,
+      accelerator: menuShortcuts.previousNote,
       click () {
         mainWindow.webContents.send('list:prior')
       }
@@ -226,7 +216,7 @@ const view = {
     },
     {
       label: 'Focus Search',
-      accelerator: menuShortcuts.focusSearch || DEFAULT_SHORTCUTS.focusSearch,
+      accelerator: menuShortcuts.focusSearch,
       click () {
         mainWindow.webContents.send('top:focus-search')
       }
